requirements.txt
streamlit==0.73.0
numpy==1.18.5
pillow
pytesseract==0.3.9
regex==2020.10.28